# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "拆镜学 API",
    "description": "拆镜学原型后端：上传参考视频 -> 拆解（镜头边界/关键帧/台词/景别/画面描述）-> 选择学习目标 -> 生成可完成的学习任务 -> 提交作品 -> 结构差异与证据化反馈。所有对外部模型的调用都发生在服务端。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "提交拆解任务（幂等：Idempotency-Key）",
        "operationId": "create_analysis_api_v1_projects__project_id__analyses_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreatedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "分析列表（按项目或状态过滤）",
        "operationId": "list_analyses_api_v1_analyses_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Analyses Api V1 Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "分析详情（分镜、解释、证据、版本）",
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/explanations": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "按学习意图生成解释（事实时间线不变）",
        "operationId": "get_explanations_api_v1_analyses__analysis_id__explanations_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "intent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Intent"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/shots": {
      "patch": {
        "tags": [
          "analyses"
        ],
        "summary": "人工修正镜头边界与标签（生成新版本）",
        "operationId": "patch_shots_api_v1_analyses__analysis_id__shots_patch",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShotsPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/versions": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "版本历史（下游任务仍引用旧版本）",
        "operationId": "list_versions_api_v1_analyses__analysis_id__versions_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Versions Api V1 Analyses  Analysis Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/learning-tasks": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "生成学习任务（模仿 / 变体 / 原创）",
        "operationId": "create_learning_task_api_v1_analyses__analysis_id__learning_tasks_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningTaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningTaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/export": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "导出 Markdown 报告（与所选版本一致）",
        "operationId": "export_analysis_api_v1_analyses__analysis_id__export_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "md",
              "title": "Format"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisCreate": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "intent": {
            "type": "string",
            "enum": [
              "shot_language",
              "narrative_rhythm"
            ],
            "title": "Intent",
            "default": "shot_language"
          },
          "generate_task": {
            "type": "boolean",
            "title": "Generate Task",
            "default": false
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "asset_id"
        ],
        "title": "AnalysisCreate"
      },
      "AnalysisOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "default": ""
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": ""
          },
          "knowledge_version": {
            "type": "string",
            "title": "Knowledge Version",
            "default": ""
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": ""
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version",
            "default": ""
          },
          "coverage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage"
          },
          "transcript": {
            "additionalProperties": true,
            "type": "object",
            "title": "Transcript"
          },
          "media": {
            "additionalProperties": true,
            "type": "object",
            "title": "Media"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "unknowns": {
            "additionalProperties": true,
            "type": "object",
            "title": "Unknowns"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status",
            "default": "unreviewed"
          },
          "edited_by_user": {
            "type": "boolean",
            "title": "Edited By User",
            "default": false
          },
          "is_demo_prebuilt": {
            "type": "boolean",
            "title": "Is Demo Prebuilt",
            "default": false
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/ShotOut"
            },
            "type": "array",
            "title": "Shots"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "project_id",
          "version",
          "kind",
          "intent",
          "status"
        ],
        "title": "AnalysisOut"
      },
...
      "ExplanationsOut": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "intent_label": {
            "type": "string",
            "title": "Intent Label"
          },
          "fact_timeline_unchanged": {
            "type": "boolean",
            "title": "Fact Timeline Unchanged"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "explanations": {
            "items": {
              "$ref": "#/components/schemas/ExplanationItem"
            },
            "type": "array",
            "title": "Explanations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "analysis_id",
          "intent",
          "intent_label",
          "fact_timeline_unchanged",
          "note",
          "explanations"
        ],
        "title": "ExplanationsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreatedOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "idempotent_replay": {
            "type": "boolean",
            "title": "Idempotent Replay",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "stage"
        ],
        "title": "JobCreatedOut"
      },
...
      "LearningTaskCreate": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "imitate",
              "variant",
              "original"
            ],
            "title": "Level",
            "default": "imitate"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "shot_language",
                  "narrative_rhythm"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          }
        },
        "type": "object",
        "title": "LearningTaskCreate"
      },
...
    }
  }
}